reports/requirements.txt
matplotlib
numpy
panda
seaborn